REFS #66 * Rename modules.
