+++ minos/networks/broker/__init__.py
@@ -5,13 +5,13 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-from .command import (
+from .commands import (
     MinosCommandBroker,
 )
-from .dispatcher import (
+from .dispatchers import (
     MinosQueueDispatcher,
 )
-from .event import (
+from .events import (
     MinosEventBroker,
 )
 from .services import (
